Remove the caseworker agent module
The /caseworker LLM-assistant portal (Skills/Summaries/Drafts/MCP servers/
LLM providers) was never linked from the site and its frontend is being
deleted (jawafdehi#drop-caseworker-agent-module). With no UI, the backend app
is dead except for three pieces, which are preserved:

- MeView (GET /api/caseworker/me): relocated verbatim into cases/api_views.py
  and mounted directly in config/urls.py. Still used by jawafdehi-mcp for chat
  identity resolution.
- /api/caseworker/auth/token/ + /refresh/: untouched (shared JWT for the
  casework review portal + poller).
- LLMService DB fallback in tag_enricher: removed. TagEnricher is now CLI-only
  and raises if no llm_client is supplied; enrich_ciaa_tags fails fast without
  --llm-base-url (pass --no-llm for rules-only).

Deletes the caseworker app, drops it from INSTALLED_APPS, and adds
cases migration 0027 to drop the orphaned caseworker_* tables (FK-safe order,
idempotent, also purges stale django_migrations rows).
