+++ cases/api_views.py
@@ -33,6 +33,12 @@ from rest_framework.throttling import AnonRateThrottle
 from rest_framework.views import APIView
 from rest_framework_simplejwt.authentication import JWTAuthentication
 
+from config.auth import (
+    JAWAFDEHI_USER_ID_HEADER,
+    SERVICE_ACCOUNT_USERNAME,
+    resolve_or_create_identity,
+)
+
 from .admin import CaseAdminForm
 from .caseworker_serializers import (
     BLOCKED_PATH_PREFIXES,
@@ -1325,3 +1331,72 @@ class OEmbedView(APIView):
             root, encoding="unicode"
         )
         return HttpResponse(xml_str, content_type="text/xml")
+
+
+class MeView(APIView):
+    """Resolve the calling chat identity to a Jawafdehi user.
+
+    Called by the jawafdehi-mcp server (GET /api/caseworker/me) using the
+    chat-jawafdehi-org service-account token plus an X-Jawafdehi-User-Id header.
+    """
+
+    authentication_classes = [TokenAuthentication]
+
+    def get(self, request):
+        if not request.user.is_authenticated:
+            return Response(
+                {"error": "Authentication required"},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
+        if not request.auth or request.auth.user.username != SERVICE_ACCOUNT_USERNAME:
+            return Response(
+                {"error": "Service account token required"},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
+        owui_user_id = (request.META.get(JAWAFDEHI_USER_ID_HEADER) or "").strip()
+        if not owui_user_id:
+            return Response(
+                {"error": "X-Jawafdehi-User-Id header is required"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+        identity = resolve_or_create_identity(owui_user_id, request)
+        if identity is None:
+            return Response(
+                {"error": f"Unknown user: {owui_user_id}"},
+                status=status.HTTP_404_NOT_FOUND,
+            )
+
+        real_user = identity.user
+
+        if real_user is None:
+            return Response(
+                {
+                    "mapped": False,
+                    "owui_user_id": identity.owui_user_id,
+                    "owui_user_name": identity.owui_user_name,
+                    "message": "Chat identity is not yet mapped to a Jawafdehi user. An admin must link this identity in the admin panel.",
+                },
+                status=status.HTTP_200_OK,
+            )
+
+        if not real_user.is_active:
+            return Response(
+                {"error": "User account is inactive"},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
+        roles = list(real_user.groups.values_list("name", flat=True))
+
+        return Response(
+            {
+                "mapped": True,
+                "roles": roles,
+                "user_id": real_user.id,
+                "username": real_user.get_username(),
+                "owui_user_id": identity.owui_user_id,
+                "owui_user_name": identity.owui_user_name,
+            }
+        )
