--- caseworker/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class CaseworkerConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "caseworker"
-    verbose_name = "Caseworker Agent"
