--- caseworker/views.py
@@ -1,272 +1,0 @@
-import logging
-import re
-
-from django.shortcuts import get_object_or_404
-from rest_framework import status, viewsets
-from rest_framework.authentication import TokenAuthentication
-from rest_framework.decorators import action
-from rest_framework.permissions import IsAuthenticated
-from rest_framework.response import Response
-from rest_framework.views import APIView
-
-from config.auth import (
-    JAWAFDEHI_USER_ID_HEADER,
-    SERVICE_ACCOUNT_USERNAME,
-    resolve_or_create_identity,
-)
-
-from .models import Draft, DraftVersion, LLMProvider, MCPServer, Skill, Summary
-from .permissions import IsAdminOrReadOnly, IsOwnerOrAdmin
-from .serializers import (
-    CurrentUserSerializer,
-    DraftSerializer,
-    DraftVersionSerializer,
-    LLMProviderSerializer,
-    MCPServerSerializer,
-    SkillSerializer,
-    SummarySerializer,
-)
-from .services import LLMService, MCPService, SummaryGenerationService
-
-logger = logging.getLogger(__name__)
-
-
-class UserViewSet(viewsets.ViewSet):
-    """Current user endpoint."""
-
-    @action(detail=False, methods=["get"], permission_classes=[IsAuthenticated])
-    def me(self, request):
-        return Response(CurrentUserSerializer(request.user).data)
-
-
-class QueryViewSet(viewsets.ViewSet):
-    """Query processing and case data retrieval."""
-
-    permission_classes = [IsAuthenticated]
-
-    @action(detail=False, methods=["post"])
-    def extract_case_number(self, request):
-        query_text = request.data.get("query", "")
-        if not query_text:
-            return Response(
-                {"error": "Query text is required"},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        match = re.search(
-            r"(\d{3}-\d{3}-\d{4}|\d{3}-CR-\d{4})", query_text, re.IGNORECASE
-        )
-        if match:
-            return Response({"case_number": match.group(1), "found": True})
-        return Response({"case_number": None, "found": False})
-
-    @action(detail=False, methods=["get"])
-    def get_case_data(self, request):
-        case_number = request.query_params.get("case_number")
-        if not case_number:
-            return Response(
-                {"error": "case_number query parameter is required"},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-        try:
-            mcp_service = MCPService()
-            case_data = mcp_service.retrieve_case_data(case_number)
-            if case_data:
-                return Response(
-                    {"case_number": case_number, "data": case_data, "found": True}
-                )
-            return Response(
-                {"error": f"Case {case_number} not found"},
-                status=status.HTTP_404_NOT_FOUND,
-            )
-        except Exception as e:
-            logger.error(f"Error retrieving case data: {e}")
-            return Response(
-                {"error": f"Failed to retrieve case data: {e}"},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            )
-
-
-class MCPServerViewSet(viewsets.ModelViewSet):
-    queryset = MCPServer.objects.all()
-    serializer_class = MCPServerSerializer
-    permission_classes = [IsAuthenticated, IsAdminOrReadOnly]
-
-    @action(detail=True, methods=["post"])
-    def test_connection(self, request, pk=None):
-        server = self.get_object()
-        is_connected = MCPService().test_connection(server)
-        server.status = "connected" if is_connected else "error"
-        server.save(update_fields=["status"])
-        return Response({"connected": is_connected})
-
-
-class SkillViewSet(viewsets.ModelViewSet):
-    queryset = Skill.objects.all()
-    serializer_class = SkillSerializer
-    permission_classes = [IsAuthenticated, IsAdminOrReadOnly]
-
-
-class SummaryViewSet(viewsets.ModelViewSet):
-    serializer_class = SummarySerializer
-    permission_classes = [IsAuthenticated, IsOwnerOrAdmin]
-
-    def get_queryset(self):
-        user = self.request.user
-        if user.is_staff:
-            return Summary.objects.all()
-        return Summary.objects.filter(user=user)
-
-    @action(detail=False, methods=["post"])
-    def generate(self, request):
-        case_number = request.data.get("case_number")
-        skill_id = request.data.get("skill_id")
-        query = request.data.get("query", "")
-
-        if not case_number or not skill_id:
-            return Response(
-                {"error": "case_number and skill_id are required"},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        try:
-            skill = get_object_or_404(Skill, id=skill_id)
-
-            summary_service = SummaryGenerationService()
-            is_valid, message = summary_service.validate_skill_prompt(skill)
-            if not is_valid:
-                return Response(
-                    {"error": f"Invalid skill prompt: {message}"},
-                    status=status.HTTP_400_BAD_REQUEST,
-                )
-
-            case_data = MCPService().retrieve_case_data(case_number)
-            content = summary_service.generate_summary(case_data, skill, query)
-
-            summary = Summary.objects.create(
-                user=request.user,
-                case_number=case_number,
-                skill=skill,
-                content=content,
-            )
-            return Response(
-                self.get_serializer(summary).data, status=status.HTTP_201_CREATED
-            )
-
-        except Exception as e:
-            logger.error(f"Error generating summary: {e}")
-            return Response(
-                {"error": "Failed to generate summary. Please try again later."},
-                status=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            )
-
-
-class DraftViewSet(viewsets.ModelViewSet):
-    serializer_class = DraftSerializer
-    permission_classes = [IsAuthenticated, IsOwnerOrAdmin]
-
-    def get_queryset(self):
-        user = self.request.user
-        if user.is_staff:
-            return Draft.objects.all()
-        return Draft.objects.filter(user=user)
-
-    def perform_create(self, serializer):
-        serializer.save(user=self.request.user)
-
-    def perform_update(self, serializer):
-        draft = self.get_object()
-        old_content = draft.content
-        serializer.save()
-        if old_content != serializer.instance.content:
-            DraftVersion.objects.create(draft=draft, content=old_content)
-
-    @action(detail=True, methods=["get"])
-    def versions(self, request, pk=None):
-        draft = self.get_object()
-        return Response(DraftVersionSerializer(draft.versions.all(), many=True).data)
-
-    @action(detail=True, methods=["post"])
-    def restore_version(self, request, pk=None):
-        draft = self.get_object()
-        version_id = request.data.get("version_id")
-        version = get_object_or_404(DraftVersion, id=version_id, draft=draft)
-        DraftVersion.objects.create(draft=draft, content=draft.content)
-        draft.content = version.content
-        draft.save()
-        return Response(self.get_serializer(draft).data)
-
-
-class LLMProviderViewSet(viewsets.ModelViewSet):
-    queryset = LLMProvider.objects.all()
-    serializer_class = LLMProviderSerializer
-    permission_classes = [IsAuthenticated, IsAdminOrReadOnly]
-
-    @action(detail=True, methods=["post"])
-    def test_connection(self, request, pk=None):
-        provider = self.get_object()
-        is_connected = LLMService().test_connection(provider)
-        return Response({"connected": is_connected})
-
-
-class MeView(APIView):
-    authentication_classes = [TokenAuthentication]
-
-    def get(self, request):
-        if not request.user.is_authenticated:
-            return Response(
-                {"error": "Authentication required"},
-                status=status.HTTP_403_FORBIDDEN,
-            )
-
-        if not request.auth or request.auth.user.username != SERVICE_ACCOUNT_USERNAME:
-            return Response(
-                {"error": "Service account token required"},
-                status=status.HTTP_403_FORBIDDEN,
-            )
-
-        owui_user_id = (request.META.get(JAWAFDEHI_USER_ID_HEADER) or "").strip()
-        if not owui_user_id:
-            return Response(
-                {"error": "X-Jawafdehi-User-Id header is required"},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-
-        identity = resolve_or_create_identity(owui_user_id, request)
-        if identity is None:
-            return Response(
-                {"error": f"Unknown user: {owui_user_id}"},
-                status=status.HTTP_404_NOT_FOUND,
-            )
-
-        real_user = identity.user
-
-        if real_user is None:
-            return Response(
-                {
-                    "mapped": False,
-                    "owui_user_id": identity.owui_user_id,
-                    "owui_user_name": identity.owui_user_name,
-                    "message": "Chat identity is not yet mapped to a Jawafdehi user. An admin must link this identity in the admin panel.",
-                },
-                status=status.HTTP_200_OK,
-            )
-
-        if not real_user.is_active:
-            return Response(
-                {"error": "User account is inactive"},
-                status=status.HTTP_403_FORBIDDEN,
-            )
-
-        roles = list(real_user.groups.values_list("name", flat=True))
-
-        return Response(
-            {
-                "mapped": True,
-                "roles": roles,
-                "user_id": real_user.id,
-                "username": real_user.get_username(),
-                "owui_user_id": identity.owui_user_id,
-                "owui_user_name": identity.owui_user_name,
-            }
-        )
