+++ config/settings.py
@@ -209,7 +209,6 @@ INSTALLED_APPS = [
     "cases",
     "nesq",
     "ngm",
-    "caseworker",
     "case_workflows",
     # Casework Review System (VOL-3): rule-centered case-quality review.
     "review",
